+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [0.3.4] - 2025-03-06
-- add minimumColumn
+- add sumColumn(), minimumColumn(), maximumColumn() and averageColumn()
 - update readme.md
 - update unit test
 - minor edits